Break connection code out into separate class
The connect entrypoint is used both for Gateway
links (jetbrains-gateway://) and internally in our own code at the end
of the wizard and for recent connections, but this will be two separate
sets of parameters (internally we have the CLI set up and pass around
the hostname while the links will only have the workspace ID or name).

So break out the code into a separate class that we can call internally
which will let us dedicate the connect entrypoint to handle the Gateway
links.  There we will set up the CLI and gather the required parameters
before calling the now-broken-out code.
